requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pymongo>=4.5.0
pyyaml>=6.0
lxml>=4.9.0
playwright>=1.40.0
matplotlib>=3.7.0

